+++ src/app/(admin)/admin/events/components/ActionDropdown/ActionDropdown.jsx
@@ -1,30 +1,16 @@
 'use client'
 
-import { useEffect, useRef, useState } from "react"
 import BtnEditUser from "../BtnEditEvent/BtnEditEvent"
 import BtnDeleteUser from "../BtnDeleteEvent/BtnDeleteEvent"
-import Link from "next/link"
+import useDropdownBlur from "@/utils/useDropDownBlur";
 
 function ActionDropdownProduct({ data }) {
-    const [toggle, SetToggle] = useState(false)
-    const menuRef = useRef()
-
-    useEffect(() => {
-        const handleBlur = (event) => {
-          if (menuRef.current && !menuRef.current.contains(event.relatedTarget)) {
-            SetToggle(false)
-          }
-        }
-        menuRef.current?.addEventListener('blur', handleBlur, true)
-        return () => {
-            menuRef.current?.removeEventListener('blur', handleBlur, true)
-        }
-    }, [])
+    const { toggle, setToggle, menuRef } = useDropdownBlur();
 
     return (
         <div className="relative" ref={menuRef}>
             <button className="px-1 py-1 text-gray-500 dark:text-gray-300 transition-colors duration-200 rounded-lg hover:bg-gray-100 hover:text-gray-500"
-                onClick={() => SetToggle(!toggle)}
+                onClick={() => setToggle(!toggle)}
             >
                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor" className="w-6 h-6">
                     <path strokeLinecap="round" strokeLinejoin="round" d="M12 6.75a.75.75 0 110-1.5.75.75 0 010 1.5zM12 12.75a.75.75 0 110-1.5.75.75 0 010 1.5zM12 18.75a.75.75 0 110-1.5.75.75 0 010 1.5z" />
